Move the Ability to the Image directory. Minor changes to match coding patterns in other files
